apps/ct_datalake/requirements.txt
openai
faiss-cpu
numpy
sentence-transformers
deep-translator
PyMuPDF
pypdf
python-docx
